# Conflicts:
#	src/main/java/com/avitam/fantasy11/core/service/NodeServiceImpl.java
#	src/main/java/com/avitam/fantasy11/core/service/UserService.java
#	src/main/java/com/avitam/fantasy11/core/service/impl/CheckoutServiceImpl.java
#	src/main/java/com/avitam/fantasy11/core/service/impl/UseServiceImpl.java
#	src/main/java/com/avitam/fantasy11/core/service/impl/UserDetailsServiceImpl.java
#	src/main/java/com/avitam/fantasy11/form/RoleForm.java
#	src/main/java/com/avitam/fantasy11/form/UserForm.java
#	src/main/java/com/avitam/fantasy11/model/BaseEntity.java
#	src/main/java/com/avitam/fantasy11/model/NodeRepository.java
#	src/main/java/com/avitam/fantasy11/model/Role.java
#	src/main/java/com/avitam/fantasy11/model/RoleRepository.java
#	src/main/java/com/avitam/fantasy11/model/User.java
#	src/main/java/com/avitam/fantasy11/model/UserRepository.java
#	src/main/java/com/avitam/fantasy11/model/VerificationToken.java
#	src/main/java/com/avitam/fantasy11/model/VerificationTokenRepository.java
#	src/main/java/com/avitam/fantasy11/suite/Fantasy11Application.java
#	src/main/java/com/avitam/fantasy11/validation/UserFormValidator.java
#	src/main/java/com/avitam/fantasy11/validation/UserValidator.java
#	src/main/java/com/avitam/fantasy11/web/controllers/AdminController.java
#	src/main/java/com/avitam/fantasy11/web/controllers/ChartReportController.java

+++ src/main/java/com/avitam/fantasy11/model/ContestJoinedRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContestJoinedRepository extends MongoRepository<ContestJoined, Integer> {
+public interface ContestJoinedRepository extends MongoRepository<ContestJoined, Long> {
 }
