+++ src/main/java/com/avitam/fantasy11/model/KYCRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KYCRepository extends MongoRepository<KYC, Integer> {
+public interface KYCRepository extends MongoRepository<KYC, Long> {
 }
