+++ src/main/java/com/avitam/fantasy11/model/LeaderBoardRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LeaderBoardRepository extends MongoRepository<LeaderBoard, Integer>{
+public interface LeaderBoardRepository extends MongoRepository<LeaderBoard, Long>{
 }
