+++ src/main/java/com/avitam/fantasy11/model/MatchTypeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MatchTypeRepository extends MongoRepository<MatchType,Integer> {
+public interface MatchTypeRepository extends MongoRepository<MatchType, Long> {
 }
