+++ src/main/java/com/avitam/fantasy11/model/TeamLineupRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamLineupRepository extends MongoRepository<TeamLineup,Integer> {
+public interface TeamLineupRepository extends MongoRepository<TeamLineup, Long> {
 }
