+++ src/main/java/com/avitam/fantasy11/model/TournamentRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TournamentRepository extends MongoRepository<Tournament,Integer> {
+public interface TournamentRepository extends MongoRepository<Tournament, Long> {
 }
